Final Commit
Ready for Review. Added in the controls and labels for the form. Added in the patterns for the controls that required them. Added in a default value for the controls that required them. Added in a placeholder for the controls that required them. Added in a submit button with an image to submit the form. Added in the CSS and Javascript files required for the page. Added in the CSS to style the client side validation by changing the background color and adding a checkmark or red x. Added in comments into the CSS files. Added in comments into the html file. Renamed the files to index.html and sb_validate.css respectively. Fixed a bug in the sb_forms.css file that would make the submit button stay on the same line as the controls.